refactor stratum methods to files
