+++ client/lib/strata/stratum/semanticZoom.js
@@ -1,0 +1,9 @@
+const stratumView = require('./view')
+
+module.exports = (stratum, space) => {
+  const viewport = space.getViewport()
+
+  if (stratum.alive) {
+    stratumView.refreshLabels(stratum, viewport)
+  }
+}
